requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=1.1.0=cpu
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py37_0
astor=0.8.0=py37_0
attrs=19.1.0=py37_1
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
cffi=1.12.3=py37h7a1dbc1_0
chardet=3.0.4=pypi_0
cloudpickle=1.2.2=py_0
colorama=0.4.1=py37_0
cuda90=1.0=0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cupy=6.0.0=py37h230ac6f_0
cycler=0.10.0=pypi_0
cytoolz=0.10.0=py37he774522_0
dask-core=2.3.0=py_0
decorator=4.4.0=py37_1
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
fastrlock=0.4=py37h6538335_0
ffmpeg=1.4=pypi_0
filelock=3.0.12=pypi_0
freetype=2.9.1=ha9979f8_1
gast=0.2.2=py37_0
grpcio=1.16.1=py37h351948d_1
h5py=2.8.0=py37hf7173ca_2
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
imageio=2.5.0=py37_0
imageio-ffmpeg=0.3.0=pypi_0
importlib-metadata=0.23=pypi_0
intel-openmp=2019.4=245
ipykernel=5.1.2=py37h39e3cac_0
ipython=7.8.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.15.1=py37_0
jinja2=2.10.1=py37_0
joblib=0.13.2=py37_0
jpeg=9b=hb83a4c4_2
jsonschema=3.0.2=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.1=py_0
jupyter_console=6.0.0=py37_0
jupyter_core=4.5.0=py_0
keras=2.2.4=pypi_0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py37_0
keras-gpu=2.2.4=0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37ha925a31_0
libopencv=3.4.2=h20b85fd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.8.0=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libsvm=3.23=pypi_0
libtiff=4.0.10=hb898794_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=pypi_0
mistune=0.8.4=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.14=py37h14836fe_0
mkl_random=1.0.2=py37h343c172_0
more-itertools=7.2.0=pypi_0
moviepy=1.0.0=pypi_0
msys2-conda-epoch=20160418=1
nbconvert=5.5.0=py_0
nbformat=4.4.0=py37_0
networkx=2.3=py_0
ninja=1.9.0=py37h74a9793_0
notebook=6.0.1=py37_0
numpy=1.16.5=py37h19fb1c0_0
numpy-base=1.16.5=py37hc3f5095_0
olefile=0.46=py37_0
opencv=3.4.2=py37h40b0b35_0
openssl=1.1.1=he774522_0
packaging=19.2=pypi_0
pandas=0.25.1=py37ha925a31_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.1=py_0
pickleshare=0.7.5=py37_0
pillow=6.1.0=py37hdc69c19_0
pip=19.2.2=py37_0
pluggy=0.13.0=pypi_0
prettytable=0.7.2=pypi_0
proglog=0.1.9=pypi_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.9=py37_0
protobuf=3.8.0=py37h33f27b4_0
psutil=5.6.7=pypi_0
py=1.8.0=pypi_0
py-opencv=3.4.2=py37hc319ecb_0
pycparser=2.19=py37_0
pygments=2.4.2=py_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.14.11=py37he774522_0
python=3.7.0=hea74fb7_0
python-dateutil=2.8.0=py37_0
pytorch=1.0.1=cpu_py37h39a92a0_0
pytz=2019.2=py_0
pywavelets=1.0.3=py37h8c2d366_1
pywinpty=0.5.5=py37_1000
pyyaml=5.1.2=py37he774522_0
pyzmq=18.1.0=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.5.5=py_0
requests=2.22.0=pypi_0
scikit-image=0.15.0=py37ha925a31_0
scikit-learn=0.21.2=py37h6288b17_0
scipy=1.1.0=pypi_0
send2trash=1.5.0=py37_0
setuptools=41.0.1=py37_0
sip=4.19.8=py37h6538335_0
six=1.12.0=py37_0
sqlite=3.29.0=he774522_0
sureal=0.4.1=pypi_0
tensorboard=1.14.0=py37he3c9ec2_0
tensorflow=1.14.0=gpu_py37h5512b17_0
tensorflow-base=1.14.0=gpu_py37h55fc52a_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py37_1
terminado=0.8.2=py37_0
testpath=0.4.2=py37_0
tk=8.6.8=hfa6e2cd_0
toml=0.10.0=pypi_0
toolz=0.10.0=py_0
torchvision=0.2.1=py_2
tornado=6.0.3=py37he774522_0
tox=3.14.0=pypi_0
tqdm=4.32.1=py_0
traitlets=4.3.2=py37_0
urllib3=1.25.3=pypi_0
vc=14.1=h0510ff6_4
virtualenv=16.7.5=pypi_0
vs2015_runtime=14.16.27012=hf0eaf9b_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=0.15.5=py_0
wheel=0.33.4=py37_0
widgetsnbextension=3.5.1=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.11.2=py37he774522_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zeromq=4.3.1=h33f27b4_3
zipp=0.6.0=pypi_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
